#106 wrong negative points count
Former-commit-id: 163aa5f51a2208586f11b46ca33e559d994137c1
Former-commit-id: 857148cea98fc596c558ca2fd633d1ea48eaa9bb
Former-commit-id: 669734f1865828bd4cd6e87b4d26afe2502bf90e [formerly c25b6688abc8c79c458eff4555dff54fab0870be]
Former-commit-id: 4015746c37a5f434fb103d73e4e4ce9d5af8958f
Former-commit-id: 29b0f320703196b298e250c3081e039e43a8548a
Former-commit-id: 4da2519ffa35739a6f2bb2610a0f9306b5402c94

+++ src/cxx_utils.cpp
@@ -54,13 +54,13 @@ int fast_countequal(NumericVector x, double t)
 // [[Rcpp::export]]
 int fast_countbelow(NumericVector x, double t)
 {
-  return std::count_if(x.begin(), x.end(), std::bind2nd(std::less<int>(), t));
+  return std::count_if(x.begin(), x.end(), std::bind2nd(std::less<double>(), t));
 }
 
 // [[Rcpp::export]]
 int fast_countover(NumericVector x, double t)
 {
-  return std::count_if(x.begin(), x.end(), std::bind2nd(std::greater<int>(), t));
+  return std::count_if(x.begin(), x.end(), std::bind2nd(std::greater<double>(), t));
 }
 
 // [[Rcpp::export]]
